DI-19262 - code refactoring
